+++ routes/zhihu/pin/utils.js
@@ -1,0 +1,48 @@
+module.exports = {
+    generateData: (data) =>
+        data.map((item) => {
+            const target = item.target ? item.target : item;
+            const pubDate = new Date(target.created * 1000).toUTCString();
+            const author = target.author.name;
+            const title = `${author}：${target.excerpt_title}`;
+            const link = `https://www.zhihu.com/pin/${target.id}`;
+            const description = target.content.reduce((description, item) => {
+                switch (item.type) {
+                    case 'text':
+                        description += `<div>${item.content}</div>`;
+                        break;
+
+                    case 'image':
+                        description += `<img referrerpolicy="no-referrer" src="${item.url.replace(/_.+\.jpg/g, '.jpg')}" />`;
+                        break;
+
+                    case 'video':
+                        description += `<video
+                    controls="controls"
+                    width="${item.playlist.hd.width}"
+                    height="${item.playlist.hd.height}"
+                    poster="${item.cover_info.thumbnail}"
+                    src="${item.playlist.hd.play_url}"
+                  >`;
+                        break;
+
+                    case 'link':
+                        description += `<div><a href="${item.url}">${item.title}</a><br><img referrerpolicy="no-referrer" src="${item.image_url}" /></div>`;
+                        break;
+
+                    default:
+                        description += '未知类型，请点击<a href="https://github.com/DIYgod/RSSHub/issues">链接</a>提交issue';
+                }
+
+                return description;
+            }, `<a href="https://www.zhihu.com${target.author.url}">${author}</a>：`);
+
+            return {
+                title,
+                description,
+                author,
+                pubDate,
+                link,
+            };
+        }),
+};
